Move disks validation from vmi admitter to pkg storage
+ Adjust existing import with proper camel case
