+++ pkg/storage/admitters/BUILD.bazel
@@ -4,6 +4,7 @@ go_test(
     name = "go_default_test",
     srcs = [
         "admit_suite_test.go",
+        "disks_test.go",
         "vm-storage-admitter_test.go",
         "vmexport_test.go",
         "vmrestore_test.go",
@@ -43,6 +44,7 @@ go_library(
     name = "go_default_library",
     srcs = [
         "data-volume-template.go",
+        "disks.go",
         "vm-storage-admitter.go",
         "vm-storage-status.go",
         "vmexport.go",
@@ -53,6 +55,7 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/storage/backend-storage:go_default_library",
+        "//pkg/util/hardware:go_default_library",
         "//pkg/util/webhooks:go_default_library",
         "//pkg/virt-config:go_default_library",
         "//staging/src/kubevirt.io/api/core:go_default_library",
@@ -66,6 +69,7 @@ go_library(
         "//vendor/k8s.io/apimachinery/pkg/api/equality:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/api/errors:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
+        "//vendor/k8s.io/apimachinery/pkg/util/validation:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/util/validation/field:go_default_library",
         "//vendor/k8s.io/client-go/tools/cache:go_default_library",
         "//vendor/kubevirt.io/containerized-data-importer-api/pkg/apis/core/v1beta1:go_default_library",
