+++ pkg/virt-api/webhooks/validating-webhook/admitters/vmi-preset-admitter.go
@@ -30,6 +30,7 @@ import (
 
 	v1 "kubevirt.io/api/core/v1"
 
+	storageadmitters "kubevirt.io/kubevirt/pkg/storage/admitters"
 	webhookutils "kubevirt.io/kubevirt/pkg/util/webhooks"
 	"kubevirt.io/kubevirt/pkg/virt-api/webhooks"
 )
@@ -76,13 +77,7 @@ func ValidateVMIPresetSpec(field *k8sfield.Path, spec *v1.VirtualMachineInstance
 		})
 	}
 
-	causes = append(causes, validateDomainPresetSpec(field.Child("domain"), spec.Domain)...)
-	return causes
-}
-
-func validateDomainPresetSpec(field *k8sfield.Path, spec *v1.DomainSpec) []metav1.StatusCause {
-	var causes []metav1.StatusCause
-	causes = append(causes, validateDevices(field.Child("devices"), &spec.Devices)...)
-	causes = append(causes, validateFirmware(field.Child("firmware"), spec.Firmware)...)
+	causes = append(causes, storageadmitters.ValidateDisks(field.Child("domain").Child("devices").Child("disks"), spec.Domain.Devices.Disks)...)
+	causes = append(causes, validateFirmware(field.Child("domain").Child("firmware"), spec.Domain.Firmware)...)
 	return causes
 }
